Add duration to morlet wavelet
For simplicity when performing the convolution, let's give a
defined duration to a wavelet. It will be roughly 0 at the
extremities (0 and duration_s) and it will be 1.0 in the
middle of the duration. This will ensure that we only ever
need a number of samples of the wavelet equal to the sample
rate of the signal * duration of the wavelet. The decay of
the exponential is controlled both by the duration itself,
and also by a factor k which is set to 2 and is not
configurable. Higher k means narrower wavelet.