Standardize morlet wavelength duration
We define the time duration of a wavelet as a time window at which
center one finds the peak of the wavelet and at the extremities a
fully decayed wavelet (i.e. ~0.0). For morlet wavelet this is
N / frequency. N depends on the epsilon we choose for 0. The
greater N is, the "smoother" the wavelet transform will be, but
more expensive to compute. In this commit, we set N to 8, which
seems a good value empirically.