Improve transform with multiple frequencies
Currently, we are using a morlet wavelet which contains about 2-3
oscillations. This produces very wide bands in frequency space, so
it cannot offer enough resolution for signals with multiple frequencies.

To improve the resolution, we can increase the number of oscillations.
This also means the convolutions will require more samples and thus the
transform will be slower (roughly linear). With 4 wave number (about 8-9
oscillations) we get very good results ona signal with 2 sine frequencies.

In addition to this, we should also scale the convolution by the number
of samples, otherwise wavelength at higher frequency with less samples,
will have their amplitude artificially lowered.