Use proper convolution for wavelet transform
The convolution used by the wavelet transform has inverted indices and
starts from 0 rather than from "-inf". This was done for convenience of
implementation, and the final result (frequency visualization) is the
same since the wavelet is essentially symmetric (not the imaginary part).
However, both for correctness and to avoid issues when implementing the
convolution through fourier transform, let's use the proper form.

Note, the result of the wavelet transform is still cropped from 0 to
signal.len()