Add inverse FFT
Inverse FFT can be calculated by exploiting the time-reversal property of
DFT. DFT(DFT(X)) is equal to N * X[0, N-1, .... 1], i.e. it scales the
signal by N and reverses the samples after the first. Then, by taking
the fourier transform of the transform, multiplying it by 1 / N, and then
time-reversing the samples after the first one, we get the original signal.

This is not the most efficient implementation, but for now it will suffice.