Implement convolution using Fourier transform
Let's add an implementation of the convolution in time using fourier.
This works as follows:
- The length of the convolution is signal_len + kernel_len - 1, rounded
  up to the nearest power of 2
- We pad with 0s the signal and the kernel to reach convolution_len
- We take the fourier transform of both signal and kernel (same size)
- We multiply each element of signal and kernel's transforms
- We compute the inverse transform

The result is the convolution in time padded with zeros up to the
nearest power of 2.

Note: for small enough inputs, a normal convolution is faster than
using fourier. With current fourier implementation, the speed-up
of fourier is noticeable only when the ratio between signal and
wavelet is 20:1. Only then, we use fourier-based convolution.