Speed-up rustftt fourier convolution
We can re-use the FftPlanner across the whole range of frequencies.
This results in 25-30% quicker execution.

In addition to this, it appears that inverse fourier transform from
rustftt doesn't apply the scaling factor (*N). While this doesn't
really change the final image (which is scaled between 0 and 1) it
does affect tests and is a bit inconsistent.