Parallelize wavelet transform
The convolutions performed for each frequency are completely independent
from each other. This can be naively parallelized, using the rayon crates.

Notes on the implementation: we require the frequencies to be in reverse
ordered in the result, but the parallel iterator makes no guarantees about
order. Hence, two additional stages are required: one to map the frequencies
to their respective indexes and one to order the results of parallel
iteration. These are both minimal O(n) operations.

With rustfft the speedup is much lower than expected (2x with 8 threads)
possibly because of heavy optimization/vectorization that can already take
advantage of multiple cores in the non-parallel case. Also, the planner
must be shared with mutex which increases overhead.